test(NumberNegativeValidator): add test to validate negative numbers

+++ src/main/java/ar/com/old/challenge_foro_hub/validators/CommentValidator.java
@@ -22,6 +22,6 @@ public class CommentValidator {
         NumberValidator validator = NumberValidatorBuilder.create()
                                             .add(NumberTypeValidator.NEGATIVE)
                                             .build();
-        validator.validate(Double.valueOf(id));
+        validator.validate(Math.toIntExact(id));
     }
 }
