+++ src/main/java/ar/com/old/challenge_foro_hub/validators/UserValidator.java
@@ -40,6 +40,6 @@ public class UserValidator {
         NumberValidator validator = NumberValidatorBuilder.create()
                                             .add(NumberTypeValidator.NEGATIVE)
                                             .build();
-        validator.validate(Double.valueOf(id));
+        validator.validate(Math.toIntExact(id));
     }
 }
