+++ src/main/java/ar/com/old/challenge_foro_hub/validators/numbers/NumberNullValidator.java
@@ -4,10 +4,10 @@ import ar.com.old.challenge_foro_hub.exceptions.validators.NumberValidationExcep
 import ar.com.old.challenge_foro_hub.exceptions.validators.ValidationException;
 import ar.com.old.challenge_foro_hub.validators.Validator;
 
-public class NumberNullValidator implements Validator<Double> {
+public class NumberNullValidator implements Validator<Integer> {
 
     @Override
-    public void validate(Double value) throws ValidationException {
+    public void validate(Integer value) throws ValidationException {
         if (value == null) {
             throw new NumberValidationException("El campo no puede ser nulo");
         }
